Backend/requirements.txt
flask
flask-cors
pillow
torch
tensorflow
opencv-python
transformers
